requirements.txt
mcp[cli]
mem0ai
uvicorn